Several improvements to test driver and some real tests
The test driver will now output errors and all outputs
for each test is folded.  This allows individual tests
to be very verbose without cluttering up the travis-ci
output screen.

There are now several "real" tests for siteserver.  They
demonstraight the usefulness of the folding by outputting
the results of GETs but these are initally collapsed
although still accessable if someone wants to see
the detail.